Use size_t so pointer arithmetic does not overflow

+++ modules/imgproc/src/median_blur.simd.hpp
@@ -663,9 +663,9 @@ medianBlur_SortNet( const Mat& _src, Mat& _dst, int m )
         size.width *= cn;
         for( i = 0; i < size.height; i++, dst += dstep )
         {
-            const T* row0 = src + std::max(i - 1, 0)*sstep;
-            const T* row1 = src + i*sstep;
-            const T* row2 = src + std::min(i + 1, size.height-1)*sstep;
+            const T* row0 = src + (size_t)std::max(i - 1, 0)*sstep;
+            const T* row1 = src + (size_t)i*sstep;
+            const T* row2 = src + (size_t)std::min(i + 1, size.height-1)*sstep;
             int limit = cn;
 
             for(j = 0;; )
@@ -750,11 +750,11 @@ medianBlur_SortNet( const Mat& _src, Mat& _dst, int m )
         for( i = 0; i < size.height; i++, dst += dstep )
         {
             const T* row[5];
-            row[0] = src + std::max(i - 2, 0)*sstep;
-            row[1] = src + std::max(i - 1, 0)*sstep;
-            row[2] = src + i*sstep;
-            row[3] = src + std::min(i + 1, size.height-1)*sstep;
-            row[4] = src + std::min(i + 2, size.height-1)*sstep;
+            row[0] = src + (size_t)std::max(i - 2, 0)*sstep;
+            row[1] = src + (size_t)std::max(i - 1, 0)*sstep;
+            row[2] = src + (size_t)i*sstep;
+            row[3] = src + (size_t)std::min(i + 1, size.height-1)*sstep;
+            row[4] = src + (size_t)std::min(i + 2, size.height-1)*sstep;
             int limit = cn*2;
 
             for(j = 0;; )
